Add server/index.js, index.html, and style.css file.
Set up the proxy server using http-proxy-middleware that connects to the navigation bar and footer service. The html file has a link tag that points to the appropriate css file and a script that that points to the appropriate bundle file. Added static images with css styles applied to them to flow with the page.